Changed exception handling in plotmon to address #352
The exception handling has moved to _safe_load_dataset. As this was
called in multiple locations in the plotmon of which only a few had
exception handling, this did not work as intended. All calls have been
modified to correctly handle the exceptions.

Extra exception types that can occur when loading data have been added
to the list of exceptions that are handled.